feat: llama stack run --providers
llama stack run --providers takes a list of providers in the format of api1=provider1,api2=provider2

this allows users to run with a simple list of providers.

given the architecture of `create_app`, this run config needs to be run to disk. use ~/.llama/distribution/providers-run/run.yaml each time for consistency

resolves #3956

Signed-off-by: Charlie Doern <cdoern@redhat.com>