chore(ci): remove unused recordings
Added a script to cleanup recordings. While doing this, moved the CI
matrix generation to a separate script so there is a single source of
truth for the matrix.

Ran the cleanup script as:

```
PYTHONPATH=. python scripts/cleanup_recordings.py
```

We can eventually put this as part of the pre-commit workflow to ensure
that the recordings are always up to date and that no stale recordings
are left in the repo.