fix: print help for list-deps when args not given
list-deps takes  positional args OR things like --providers

the issue with this, is that these args need to be optional since by nature, one or the other can be specified.

add a check to list-deps that checks if not args.providers and not args.config. If this is true, help is printed and we exit.

resolves #4075

Signed-off-by: Charlie Doern <cdoern@redhat.com>