fix: Query default values can't be set in Annotated
The error is that Query default values can't be set in Annotated; they
must be set with = in the function signature.
See the error:

```
The error is that Query default values can't be set in Annotated; they
must be set with = in the function signature. Searching for where
include_embeddings is defined:
```

Signed-off-by: Sébastien Han <seb@redhat.com>